Add competences' dynamic gestion + i18n + few code modifications
The competences are now brought from the json file. Once read, they're
published into the table dynamically.
The category of each competences are set as a "title" of each section
(could be cut in different tables afterward, and used to show/hide the
whole section).
The requirements are now set so we can then add color to competences
that doesn't have the proper requirements.
Also added i18n, in order to manage only keywords, and then benefit of
the i18n file to recover the names.
Changes some typo/misswriten elements.
Added some CSS style.
Change most scope of variables (var to let)